+++ annotationprocessortesting-kt/src/main/java/com/tickaroo/tikxml/regressiontests/cdata_not_filling_buffer/AudioXml.kt
@@ -1,4 +1,4 @@
-package com.tickaroo.tikxml.regressiontests.unterminatedcomment
+package com.tickaroo.tikxml.regressiontests.cdata_not_filling_buffer
 
 import com.tickaroo.tikxml.annotation.Attribute
 import com.tickaroo.tikxml.annotation.Xml
