+++ annotationprocessortesting-kt/src/test/java/com/tickaroo/tikxml/regressiontests/cdata_not_filling_buffer/CDATA_BufferNotFilledTest.kt
@@ -1,4 +1,4 @@
-package com.tickaroo.tikxml.regressiontests.unterminatedcomment
+package com.tickaroo.tikxml.regressiontests.cdata_not_filling_buffer
 
 import com.tickaroo.tikxml.TestUtils
 import com.tickaroo.tikxml.TikXml
